requirements.txt
Package               Version
--------------------- -------
asgiref               3.7.2
Django                5.0
jedi                  0.17.2
mysqlclient           2.2.1
parso                 0.7.1
pip                   23.3.1
pluggy                1.3.0
python-jsonrpc-server 0.4.0
setuptools            68.2.2
sqlparse              0.4.4
typing_extensions     4.9.0
tzdata                2023.3
ujson                 5.9.0
wheel                 0.41.3
djangorestframework   3.14.0 
pytz                  2023.3.post1
django-filter         23.5
drf-nested-routers    0.93.5
